fix: ajouté vérification de la valeur de l'octet

+++ src/mac-ip/ip.h
@@ -18,8 +18,8 @@ typedef struct ip ip;
 ip* ip_init();
 ERREUR_IP ip_deinit(ip* ip);
 
-ERREUR_IP ip_set_adresse(ip* ip, uint8_t octet1, uint8_t octet2, uint8_t octet3, uint8_t octet4);
-ERREUR_IP ip_set_octet_adresse(ip* ip, uint8_t octet, size_t index);
+ERREUR_IP ip_set_adresse(ip* ip, int octet1, int octet2, int octet3, int octet4);
+ERREUR_IP ip_set_octet_adresse(ip* ip, int octet, size_t index);
 ERREUR_IP ip_set_cidr(ip* ip , uint8_t cidr);
 ERREUR_IP ip_get_string(ip* ip, char* str);
